+++ src/RcsbFvUI/Components/ItemComponent.tsx
@@ -9,7 +9,10 @@ export class ItemComponent extends React.Component<React.HTMLAttributes<any>,nul
             className={"user-select-none d-inline-block text-white text-center bg-primary bg-gradient rounded bg-opacity-75 "+this.props.className}
             style={{
                 ...this.props.style,
-                width:120
+                width:120,
+                MozUserSelect:"none",
+                WebkitUserSelect:"none",
+                msUserSelect:"none"
             }}
         />);
     }